+++ apps/forge-web/src/layouts/dashboard/sidebar.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
+
 import { withRouter, Link } from 'react-router-dom';
 
 import Icon8 from '../../components/icon8';