+++ apps/forge-web/src/pages/explorer/account.js
@@ -2,12 +2,13 @@ import React from 'react';
 import styled from 'styled-components';
 import PropTypes from 'prop-types';
 import { withRouter } from 'react-router-dom';
+import { arc } from '@arcblock/forge-util';
 
-import Typography from '@material-ui/core/Typography';
 import CircularProgress from '@material-ui/core/CircularProgress';
 
 import Page from '../../components/page';
 import Layout from '../../layouts/page';
+import SummaryHeader from './components/summary_header';
 import withI18n from '../../components/withI18n';
 import withRoot from '../../components/withRoot';
 
@@ -34,16 +35,26 @@ class AccountDetail extends Page {
 
   render() {
     const { loading, account } = this.state;
-    const { address } = this.props.match.params;
     return (
       <Layout title="Account" cookies={this.cookies}>
         <Container>
-          <Typography component="h3">Account detail for {address}: </Typography>
           {loading && <CircularProgress />}
           {account && (
-            <pre>
-              <code>{JSON.stringify(account, true, '  ')}</code>
-            </pre>
+            <React.Fragment>
+              <SummaryHeader
+                type={account.address}
+                title={account.moniker}
+                badge={arc.fromArc(account.balance)}
+                badgeTip="Balance"
+                meta={[
+                  { key: 'First Seen', value: account.context.genesisTime },
+                  { key: 'Last Seen', value: account.context.renaissanceTime },
+                ]}
+              />
+              <pre>
+                <code>{JSON.stringify(account, true, '  ')}</code>
+              </pre>
+            </React.Fragment>
           )}
         </Container>
       </Layout>
@@ -53,13 +64,15 @@ class AccountDetail extends Page {
   async loadStatus() {
     const { address } = this.props.match.params;
     this.setState({ loading: true });
-    const { state } = await forge.getAccountState({ address: address.toLowerCase() });
+    const { state } = await forge.getAccountState({ address });
     this.setState({ loading: false, account: state });
   }
 }
 
 const Container = styled.div`
-  padding: ${props => props.theme.spacing.unit * 3}px;
+  padding: ${props => props.theme.spacing.unit * 6}px 8%;
+  width: auto;
+  max-width: 1280px;
 `;
 
 export default withRoot(withI18n(withRouter(AccountDetail)));