+++ apps/forge-web/src/pages/explorer/txs.js
@@ -1,13 +1,14 @@
 import React from 'react';
 import styled from 'styled-components';
 import PropTypes from 'prop-types';
+import qs from 'querystring';
 import { withRouter } from 'react-router-dom';
 
 import CircularProgress from '@material-ui/core/CircularProgress';
 
 import TxCard from './components/tx_card/index';
 import Pagination from './components/pagination';
-import SearchBox from './components/search_box';
+import FilterStrip from './components/filter/strip';
 import SummaryHeader from './components/summary_header';
 
 import Page from '../../components/page';
@@ -16,7 +17,7 @@ import withI18n from '../../components/withI18n';
 import withRoot from '../../components/withRoot';
 
 import forge from '../../libs/forge';
-import { parseQuery } from '../../libs/util';
+import { parseQuery, fromTypeUrl, toTypeUrl } from '../../libs/util';
 
 class TransactionList extends Page {
   static propTypes = {
@@ -28,14 +29,17 @@ class TransactionList extends Page {
   constructor(props) {
     super(props);
     const params = parseQuery(window.location.search);
+    const selected = params.filter ? params.filter.split(',').map(x => toTypeUrl(x)) : [];
 
     this.state = {
       loading: false,
+      loadingTxs: false,
       txs: null,
       chainInfo: null,
       currentPage: Number(params.page) || 1,
       pageSize: 50,
       pageParam: { next: false, cursor: null },
+      selectedTxs: selected,
     };
 
     this.loadTransactions = this.loadTransactions.bind(this);
@@ -46,10 +50,8 @@ class TransactionList extends Page {
     this.loadChainInfo(this.loadTransactions);
   }
 
-  // TODO: add search feature
-  // TODO: add filter feature
   render() {
-    const { loading, chainInfo, txs, currentPage, pageSize } = this.state;
+    const { loading, loadingTxs, chainInfo, txs, currentPage, pageSize, selectedTxs } = this.state;
     return (
       <Layout title="Transactions" cookies={this.cookies}>
         <Container>
@@ -66,8 +68,16 @@ class TransactionList extends Page {
               ]}
             />
           )}
-          {chainInfo && <SearchBox />}
-          {txs && (
+          {chainInfo && (
+            <FilterStrip
+              showFilter={true}
+              supportedTxs={chainInfo.supportedTxs.sort()}
+              selectedTxs={selectedTxs}
+              onApplyFilter={this.onApplyFilter}
+            />
+          )}
+          {loadingTxs && <CircularProgress />}
+          {!loadingTxs && txs && (
             <div className="txs">
               {txs.map(x => (
                 <TxCard key={x.hash} tx={x} />
@@ -88,39 +98,61 @@ class TransactionList extends Page {
     );
   }
 
+  onApplyFilter = selectedTxs => {
+    this.setState({ selectedTxs }, () => {
+      this.loadTransactions();
+    });
+  };
+
   onChangePage(page) {
     if (this.state.currentPage === page) {
       return;
     }
 
-    window.location.href = `/node/explorer/txs?page=${page}`;
+    const { selectedTxs } = this.state;
+    window.location.href = `/node/explorer/txs?${qs.stringify({
+      page,
+      filter: selectedTxs.map(x => fromTypeUrl(x, false)).join(','),
+    })}`;
   }
 
   async loadChainInfo(done) {
     this.setState({ loading: true });
     const { info: chainInfo } = await forge.getChainInfo();
-    this.setState({ loading: false, chainInfo }, done);
+    const { selectedTxs } = this.state;
+    this.setState(
+      {
+        loading: false,
+        chainInfo,
+        selectedTxs: selectedTxs.length ? selectedTxs : chainInfo.supportedTxs,
+      },
+      done
+    );
   }
 
   async loadTransactions() {
-    this.setState({ txs: [] });
-    const { pageSize, pageParam } = this.state;
+    this.setState({ txs: null, loadingTxs: true });
+    const { pageSize, pageParam, selectedTxs } = this.state;
 
-    const params = { paging: { size: pageSize } };
+    const params = {
+      paging: { size: pageSize },
+      typeFilter: { types: selectedTxs.map(x => fromTypeUrl(x, false)) },
+    };
     if (pageParam && pageParam.next && pageParam.cursor) {
       params.paging.cursor = pageParam.cursor;
     }
     const { transactions: txs, page } = await forge.listTransactions(params);
 
-    this.setState({ txs, page });
+    this.setState({ txs, page, loadingTxs: false });
   }
 }
 
 // TODO: make this responsive
 const Container = styled.div`
-  padding: ${props => props.theme.spacing.unit * 6}px ${props => props.theme.spacing.unit * 15}px;
+  padding: ${props => props.theme.spacing.unit * 6}px 8%;
   width: auto;
   max-width: 1280px;
+  box-sizing: border-box;
 
   .pagination {
     margin-top: 60px;